+++ lib/_core/constants/move.dart
@@ -11,7 +11,7 @@ class Move {
   // TODO : 서버 측 엔드포인트 확인 -> 변경하기
 
   static String NowBookListPage = "/nowBooks";
-  static String BookDetailPage = "/bookDetail/1";
+  static String BookDetailPage = "/bookDetail";
   static String PostDetailPage = "/postDetail";
   static String PostWritePage = "/postWrite";
   static String JoinIdEmailPage = "/joinIdEmail";
@@ -24,7 +24,7 @@ Map<String, Widget Function(BuildContext)> getRouters() {
     // 예시
     Move.NowBookListPage: (context) => const NowBookListPage(),
     Move.BookDetailPage: (context) => const BookDetailPage(bookId: 1),
-    Move.PostDetailPage: (context) => const PostDetailPage(),
+    Move.PostDetailPage: (context) => const PostDetailPage(postId: 1),
     Move.PostWritePage: (context) => PostWritePage(),
     Move.JoinIdEmailPage: (context) => const JoinIdEmailPage(),
     Move.MainSplashPage: (context) => const MainSplashPage(),
